+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.14.0] - 2021-05-31
+### Added
+- support for history states (not supported in generator yet)
+### Fixed
+- handle scenario when initial state of HSM has substates (from now on initialize() API must be called after defining HSM structure)
+
 ## [0.13.4] - 2021-05-24
 ### Added
 - HSMBUILD_DEBUGGING build option